# Refactor imresize using Interpolations