--- a/REQUIRE
+++ b/REQUIRE
@@ -1,6 +1,6 @@
 julia 0.5
 ImageCore 0.1.2
-CoordinateTransformations
+CoordinateTransformations 0.4.0
 Interpolations 0.3.7
 OffsetArrays
 StaticArrays