On my plane ride back from a work-related trip, I experimented with writing an `imresize` function using interpolations. The best I could come up with was this approach. It seems to work quite well for upscaling an Image, but when downscaling it, it is more of a sub-sampling than anything, which may not be the best way to do it.

Also, while playing around with Interpolations and Extrapolation (which are both an `AbstractArray` which trigger ImageInTerminal rendering), I noticed that restrict triggered some exceptions. For example one cant to `A[1]` if `A` is an interpolation