@Evizero 2017-02-18
> What's the performance like?

Have not benchmarked yet. To be honest I was unsure if this is even close to what you had in mind