@timholy 2017-02-18
It's basically spot on. Nice work. My only concern is the inner loop, it seems possible we'd need to optimize it to get rid of the splat. But it's possible that the compiler will do all the hard work for us (which would be much better than doing it by hand).