@Evizero 2017-02-18
Understood. I will look into performance