@Evizero 2017-02-18
To get a little glimpse of where I am at I tried the following little micro benchmark as reference for the status quo 

current master before my refactor:

```julia
julia> using Colors, ImageTransformations, BenchmarkTools

julia> org = rand(RGB, 20, 20);

julia> dst = similar(org, (1000,1000));

julia> @benchmark ImageTransformations.imresize!($dst, $org)
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     9.838 ms (0.00% GC)
  median time:      9.869 ms (0.00% GC)
  mean time:        9.873 ms (0.00% GC)
  maximum time:     10.129 ms (0.00% GC)
  --------------
  samples:          506
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
```

This PR currently without additional optimization:

```julia
julia> @benchmark ImageTransformations.imresize!($dst, $org)
BenchmarkTools.Trial: 
  memory estimate:  9.59 KiB
  allocs estimate:  5
  --------------
  minimum time:     17.118 ms (0.00% GC)
  median time:      17.126 ms (0.00% GC)
  mean time:        17.153 ms (0.00% GC)
  maximum time:     18.202 ms (0.00% GC)
  --------------
  samples:          292
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
```