@Evizero 2017-02-18
The 0.6 test issues on travis are related to https://github.com/JuliaArrays/StaticArrays.jl/issues/106