@Evizero 2017-02-18
Concerning benchmark. This implementation using a generated function yields the same regression as the current state of the PR

```julia
@generated function imresize!{T,S,N}(resized::AbstractArray{T,N}, original::AbstractInterpolation{S,N})
    quote
        org_size = size(original)
        dst_size = size(resized)
        @inbounds @nexprs $N i->(scale_i = org_size[i] / dst_size[i])
        @inbounds for I in CartesianRange(size(resized))
            I_d = I.I
            @nexprs $N i->(O_i = I_d[i] * scale_i)
            resized[I] = @nref $N original i->O_i
        end
        resized
    end
end
```

I don't think the tuple splatting is the issue. I am guessing the overhead is from interpolations maybe?