@timholy 2017-02-20
OK, let's go with the simple version and then figure out how to improve Interpolations performance.

With regards to sub-sampling etc: I think when making a smaller image we're going to first have to do some filtering. We could either use ImageFiltering, or perhaps one could just `restrict` until it's smaller and then interpolate up? I'm sure there's literature on this, but I confess I haven't read it.

This seems fine to me. I'd say merge whenever you feel ready. To me it looks like the tests bork before we get to the `imresize` tests, so just make sure those work properly and we can fix other issues separately if you prefer.