@Evizero 2017-02-23
Thanks for your comments and advice so far. I am looking forward to further contribute to this package