@timholy 2017-02-24
I think this is an important area for future growth, so I'm very glad to hear of your interest in pitching in! Right now my main focus is on getting the whole Images suite passing on 0.6, and in preparing the way for GSoC, so it's great to have people driving things forward in more "fundamental" respects.